fixed rows datatype

+++ vzReader/models/mySQLRawDataModel.go
@@ -13,6 +13,6 @@ type MySQLRawDataModel struct {
 	SpanId      string  `json:"span_id"`
 	WorkloadIds string  `json:"workload_ids"`
 	IsTruncated bool    `json:"is_truncated"`
-	Rows        int     `json:"rows"`
+	Rows        uint64  `json:"rows"`
 	Time        uint64  `json:"time"`
 }
